blockchain-nfc-attendance/backend/requirements.txt
setuptools>=65.0.0
Flask==2.3.0
web3==6.8.0
python-dotenv==1.0.0
flask-cors==4.0.0
pynfc==0.1.1
requests>=2.28.0
